+++ scripts/main.py
@@ -1,7 +1,0 @@
-from htekTestTools.htekTestTools import *
-from data.usr_data import *
-
-
-result = basic_call(phone_1, phone_2)
-print(result)
-# transfer_flow(phone_1, phone_2, phone_3)